requirements.txt
black~=23.3.0
pylint~=2.17.2
filelock~=3.11.0
numpy~=1.23.5
pandas~=2.0.0
tensorflow-cpu~=2.12.*
